// frontend/src/app/models/korisnik.ts
export class Korisnik{
    korisnicko_ime: string = "";
    lozinka: string = "";
    bezbedonosno_pitanje: string = "";
    odgovor: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    kontakt_telefon: number = 0;
    email: string = "";
    profilna_slika: string = "";
    tip_skole: number = 0;
    trenutni_razred: number = 0;
    greska: string = "";
    tip: string = "";
}